+++ test/SemanticObjectAnonymousSub.php
@@ -5,7 +5,7 @@ namespace VirtualAssembly\Semantizer;
 class SemanticObjectAnonymousSub extends SemanticObjectAnonymous {
 
     public function __construct(Semantizer $semantizer) {
-        parent::__construct($semantizer, "foaf:testType");
+        parent::__construct(semantizer: $semantizer, semanticType: "foaf:testType");
     }
 
 }